fix multi definition bug
