+++ include/imu_processing.h
@@ -18,25 +18,97 @@ namespace fast_lio {
 
 constexpr int MAX_INI_COUNT = 20;
 
-bool time_list(const PointType &x, const PointType &y) { return (x.curvature < y.curvature); };
-
 /// IMU Process and undistortion
 class ImuProcess {
    public:
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
-    ImuProcess();
-    ~ImuProcess();
+    ImuProcess() : b_first_frame_(true), imu_need_init_(true) {
+        init_iter_num_ = 1;
+        Q_ = process_noise_cov();
+        cov_acc_ = common::V3D(0.1, 0.1, 0.1);
+        cov_gyr_ = common::V3D(0.1, 0.1, 0.1);
+        cov_bias_gyr_ = common::V3D(0.0001, 0.0001, 0.0001);
+        cov_bias_acc_ = common::V3D(0.0001, 0.0001, 0.0001);
+        mean_acc_ = common::V3D(0, 0, -1.0);
+        mean_gyr_ = common::V3D(0, 0, 0);
+        angvel_last_ = common::Zero3d;
+        Lidar_T_wrt_IMU_ = common::Zero3d;
+        Lidar_R_wrt_IMU_ = common::Eye3d;
+        last_imu_.reset(new sensor_msgs::msg::Imu());
+    }
+    ~ImuProcess(){}
 
-    void Reset();
-    void SetExtrinsic(const common::V3D &transl, const common::M3D &rot);
-    Sophus::SE3d GetExtrinsic();
-    void SetGyrCov(const common::V3D &scaler);
-    void SetAccCov(const common::V3D &scaler);
-    void SetGyrBiasCov(const common::V3D &b_g);
-    void SetAccBiasCov(const common::V3D &b_a);
+    static bool time_list(const PointType &x, const PointType &y) { return (x.curvature < y.curvature); };
+
+    void Reset(){
+        mean_acc_ = common::V3D(0, 0, -1.0);
+        mean_gyr_ = common::V3D(0, 0, 0);
+        angvel_last_ = common::Zero3d;
+        imu_need_init_ = true;
+        init_iter_num_ = 1;
+        v_imu_.clear();
+        IMUpose_.clear();
+        last_imu_.reset(new sensor_msgs::msg::Imu());
+        cur_pcl_un_.reset(new PointCloudType());
+    }
+
+    void SetExtrinsic(const common::V3D &transl, const common::M3D &rot){
+        Lidar_T_wrt_IMU_ = transl;
+        Lidar_R_wrt_IMU_ = rot;
+    }
+
+    Sophus::SE3d GetExtrinsic(){
+        Sophus::SE3d extrinsicImuToLidar{ Lidar_R_wrt_IMU_, Lidar_T_wrt_IMU_ };
+        return extrinsicImuToLidar;
+    }
+
+    void SetGyrCov(const common::V3D &scaler){
+        cov_gyr_scale_ = scaler;
+    }
+    void SetAccCov(const common::V3D &scaler){
+        cov_acc_scale_ = scaler;
+    }
+    void SetGyrBiasCov(const common::V3D &b_g){
+        cov_bias_gyr_ = b_g;
+    }
+    void SetAccBiasCov(const common::V3D &b_a){
+        cov_bias_acc_ = b_a;
+    }
     void Process(const common::MeasureGroup &meas, esekfom::esekf<state_ikfom, 12, input_ikfom> &kf_state,
-                 PointCloudType::Ptr pcl_un_);
+                 PointCloudType::Ptr cur_pcl_un_){
+        if (meas.imu_.empty()) {
+            return;
+        }
+
+        if (imu_need_init_) {
+            /// The very first lidar frame
+            IMUInit(meas, kf_state, init_iter_num_);
+
+            imu_need_init_ = true;
+
+            last_imu_ = meas.imu_.back();
+
+            state_ikfom imu_state = kf_state.get_x();
+            if (init_iter_num_ > MAX_INI_COUNT) {
+                cov_acc_ *= pow(common::G_m_s2 / mean_acc_.norm(), 2);
+                imu_need_init_ = false;
+
+                cov_acc_ = cov_acc_scale_;
+                cov_gyr_ = cov_gyr_scale_;
+                LOG(WARNING) << "mean Acc: " << mean_acc_.transpose();
+                LOG(WARNING) << "cov Acc: " << cov_acc_.transpose();
+                LOG(WARNING) << "cov Gyro: " << cov_gyr_.transpose();
+                LOG(INFO) << "IMU Initial Done";
+                // fout_imu_.open(common::DEBUG_FILE_DIR("imu_.txt"), std::ios::out);  // ! maybe modify later
+            }
+
+            return;
+        }
+
+        Timer::Evaluate([&, this]() { UndistortPcl(meas, kf_state, *cur_pcl_un_); }, "Undistort Pcl");
+
+    }
 
     std::ofstream fout_imu_;
     Eigen::Matrix<double, 12, 12> Q_;
@@ -48,9 +120,209 @@ class ImuProcess {
     common::V3D cov_bias_acc_;
 
    private:
-    void IMUInit(const common::MeasureGroup &meas, esekfom::esekf<state_ikfom, 12, input_ikfom> &kf_state, int &N);
+    void IMUInit(const common::MeasureGroup &meas, esekfom::esekf<state_ikfom, 12, input_ikfom> &kf_state, int &N)
+    {
+        /** 1. initializing the gravity_, gyro bias, acc and gyro covariance
+        ** 2. normalize the acceleration measurenments to unit gravity_ **/
+
+        common::V3D cur_acc, cur_gyr;
+
+        if (b_first_frame_) {
+            Reset();
+            N = 1;
+            b_first_frame_ = false;
+            const auto &imu_acc = meas.imu_.front()->linear_acceleration;
+            const auto &gyr_acc = meas.imu_.front()->angular_velocity;
+            mean_acc_ << imu_acc.x, imu_acc.y, imu_acc.z;
+            mean_gyr_ << gyr_acc.x, gyr_acc.y, gyr_acc.z;
+        }
+
+        for (const auto &imu : meas.imu_) {
+            const auto &imu_acc = imu->linear_acceleration;
+            const auto &gyr_acc = imu->angular_velocity;
+            cur_acc << imu_acc.x, imu_acc.y, imu_acc.z;
+            cur_gyr << gyr_acc.x, gyr_acc.y, gyr_acc.z;
+
+            mean_acc_ += (cur_acc - mean_acc_) / N;
+            mean_gyr_ += (cur_gyr - mean_gyr_) / N;
+
+            cov_acc_ =
+                cov_acc_ * (N - 1.0) / N + (cur_acc - mean_acc_).cwiseProduct(cur_acc - mean_acc_) * (N - 1.0) / (N * N);
+            cov_gyr_ =
+                cov_gyr_ * (N - 1.0) / N + (cur_gyr - mean_gyr_).cwiseProduct(cur_gyr - mean_gyr_) * (N - 1.0) / (N * N);
+
+            N++;
+        }
+        state_ikfom init_state = kf_state.get_x();
+        init_state.grav = S2(-mean_acc_ / mean_acc_.norm() * common::G_m_s2);
+
+        init_state.bg = mean_gyr_;
+        init_state.offset_T_L_I = Lidar_T_wrt_IMU_;
+        init_state.offset_R_L_I = Lidar_R_wrt_IMU_;
+        kf_state.change_x(init_state);
+
+        esekfom::esekf<state_ikfom, 12, input_ikfom>::cov init_P = kf_state.get_P();
+        init_P.setIdentity();
+        init_P(6, 6) = init_P(7, 7) = init_P(8, 8) = 0.00001;
+        init_P(9, 9) = init_P(10, 10) = init_P(11, 11) = 0.00001;
+        init_P(15, 15) = init_P(16, 16) = init_P(17, 17) = 0.0001;
+        init_P(18, 18) = init_P(19, 19) = init_P(20, 20) = 0.001;
+        init_P(21, 21) = init_P(22, 22) = 0.00001;
+        kf_state.change_P(init_P);
+        last_imu_ = meas.imu_.back();
+    }
+    
     void UndistortPcl(const common::MeasureGroup &meas, esekfom::esekf<state_ikfom, 12, input_ikfom> &kf_state,
-                      PointCloudType &pcl_out);
+                      PointCloudType &pcl_out){
+        // Step 1.get head and tail time of imu and lidar
+        auto v_imu = meas.imu_;
+        v_imu.push_front(last_imu_);    // add the imu_ of the last frame-tail to the of current frame-head
+        const double &imu_beg_time = common::FromRosTime(v_imu.front()->header.stamp);
+        const double &imu_end_time = common::FromRosTime(v_imu.back()->header.stamp);
+        const double &pcl_beg_time = meas.lidar_bag_time_;
+        const double &pcl_end_time = meas.lidar_end_time_;
+
+        // Step 2.sort point clouds by offset time
+        pcl_out = *(meas.lidar_);
+        sort(pcl_out.points.begin(), pcl_out.points.end(), time_list);
+        // LOG(WARNING) << "points size: " << pcl_out.points.size();
+        // Step 3.Initialize tail IMU pose of last frame
+        state_ikfom imu_state = kf_state.get_x();
+        IMUpose_.clear();
+        IMUpose_.push_back(common::set_pose6d(0.0, acc_s_last_, angvel_last_, imu_state.vel, imu_state.pos,
+                                            imu_state.rot.toRotationMatrix()));
+
+        
+        common::V3D angvel_avr, acc_avr, acc_imu, vel_imu, pos_imu;
+        common::M3D R_imu;
+
+        double dt = 0;
+
+        input_ikfom in;
+        // Step 4.go through all imu vector， forward propagation at each imu_ date
+        for (auto it_imu = v_imu.begin(); it_imu < (v_imu.end() - 1); it_imu++) {
+            auto &&head = *(it_imu);
+            auto &&tail = *(it_imu + 1);
+            // Step 4.1 insure: last lidar end time << tail time
+            if ( common::FromRosTime(tail->header.stamp) < last_lidar_end_time_) {
+                continue;
+            }
+
+            // Step 4.2 get ave imu date
+            angvel_avr << 0.5 * (head->angular_velocity.x + tail->angular_velocity.x),
+                0.5 * (head->angular_velocity.y + tail->angular_velocity.y),
+                0.5 * (head->angular_velocity.z + tail->angular_velocity.z);
+            acc_avr << 0.5 * (head->linear_acceleration.x + tail->linear_acceleration.x),
+                0.5 * (head->linear_acceleration.y + tail->linear_acceleration.y),
+                0.5 * (head->linear_acceleration.z + tail->linear_acceleration.z);
+            
+            // Step 4.3 acc normalize to G_m_s2 length
+            acc_avr = acc_avr * common::G_m_s2 / mean_acc_.norm();  // - state_inout.ba;
+
+            // Step 4.4 get dt
+            if ( common::FromRosTime( head->header.stamp ) < last_lidar_end_time_) {
+                dt = common::FromRosTime( tail->header.stamp ) - last_lidar_end_time_;
+            } else {
+                dt = common::FromRosTime( tail->header.stamp ) - common::FromRosTime( head->header.stamp );
+            }
+
+            // Step 4.5 forward propagation this imu date
+            in.acc = acc_avr;
+            in.gyro = angvel_avr;
+            Q_.block<3, 3>(0, 0).diagonal() = cov_gyr_;
+            Q_.block<3, 3>(3, 3).diagonal() = cov_acc_;
+            Q_.block<3, 3>(6, 6).diagonal() = cov_bias_gyr_;
+            Q_.block<3, 3>(9, 9).diagonal() = cov_bias_acc_;
+            kf_state.predict(dt, Q_, in);
+
+            // Step 4.6 save the poses at each IMU measurements( in imu_world frame ) by Step4.5
+            imu_state = kf_state.get_x();
+            angvel_last_ = angvel_avr - imu_state.bg;
+            acc_s_last_ = imu_state.rot * (acc_avr - imu_state.ba);
+            for (int i = 0; i < 3; i++) {
+                acc_s_last_[i] += imu_state.grav[i];
+            }
+
+            double &&offs_t = common::FromRosTime( tail->header.stamp ) - pcl_beg_time;
+            // simple set off_t、acc、ang、vel、trans、rot
+            IMUpose_.emplace_back(common::set_pose6d(offs_t, acc_s_last_, angvel_last_, imu_state.vel, imu_state.pos,
+                                                    imu_state.rot.toRotationMatrix()));
+        }
+
+        // Step 5.forward to tail lidar point in cur frame
+        double note = pcl_end_time > imu_end_time ? 1.0 : -1.0;
+        dt = note * (pcl_end_time - imu_end_time);
+        kf_state.predict(dt, Q_, in);
+
+        imu_state = kf_state.get_x();
+        last_imu_ = meas.imu_.back();
+        last_lidar_end_time_ = pcl_end_time;
+
+        // Step 6.undistort each lidar point (backward propagation)
+        if (pcl_out.points.empty()) {
+            return;
+        }
+        auto it_pcl = pcl_out.points.end() - 1; // tail point
+        // LOG(WARNING) << "orgin itPcl: " << it_pcl->x << " " << it_pcl->y << " " << it_pcl->z;
+        // backward propagation imu pose
+        for (auto it_kp = IMUpose_.end() - 1; it_kp != IMUpose_.begin(); it_kp--) {
+            auto head = it_kp - 1;
+            auto tail = it_kp;
+            // R_imu = common::MatFromArray(head->rot);
+            // // LOG_FIRST_N(INFO,5) << "head->rot:" << std::endl << head->rot;
+            // // LOG_FIRST_N(INFO,5) << "R_imu:" << std::endl << R_imu;
+            // vel_imu = common::VecFromArray(head->vel);
+            // pos_imu = common::VecFromArray(head->pos);
+            // acc_imu = common::VecFromArray(tail->acc);
+            // angvel_avr = common::VecFromArray(tail->gyr);
+
+            R_imu = head->rot;
+            vel_imu = head->vel;
+            pos_imu = head->pos;
+            acc_imu = tail->acc;
+            angvel_avr = tail->gyr;
+
+            // LOG(INFO) << "imu pose vector offset time: " << head->offset_time;
+            // go through points between head and tail imu pose
+            for (; it_pcl->curvature / double(1000) > head->offset_time; it_pcl--) {
+                dt = it_pcl->curvature / double(1000) - head->offset_time;
+                // LOG(WARNING) << "dt: " << dt;
+                /* ! Transform to the 'end' frame, using only the rotation
+                * Note: Compensation direction is INVERSE of Frame's moving direction
+                * So if we want to compensate a point at timestamp-i to the frame-e
+                * p_compensate = R_imu_e ^ T * (R_i * P_i + T_ei) where T_ei is represented in global frame */
+                
+                // this point time rot in imu_world
+                common::M3D R_i(R_imu * Exp(angvel_avr, dt));
+                // this point time translate in lidar_body frame
+                common::V3D P_i(it_pcl->x, it_pcl->y, it_pcl->z);
+                // delta translate of (end time pose) to (this point time) in imu_world frame
+                common::V3D T_ei(pos_imu + vel_imu * dt + 0.5 * acc_imu * dt * dt - imu_state.pos);
+                // 
+                common::V3D p_compensate =
+                    /*lidar point pose in tail lidar_body frame without rotation distort*/
+                    imu_state.offset_R_L_I.conjugate() *
+                    /*lidar point pose in tail imu_body frame without rotation distort*/
+                    (imu_state.rot.conjugate() * 
+                        /*lidar point pose in imu_world frame with distort*/
+                        (R_i * 
+                            /*lidar point pose in imu_body frame*/
+                            (imu_state.offset_R_L_I * P_i + imu_state.offset_T_L_I)
+                        + T_ei) -
+                    imu_state.offset_T_L_I);  // not accurate!
+
+                // save Undistorted points and their rotation
+                it_pcl->x = p_compensate(0);
+                it_pcl->y = p_compensate(1);
+                it_pcl->z = p_compensate(2);
+                // LOG(WARNING) << "itPcl: " << it_pcl->x << " " << it_pcl->y << " " << it_pcl->z;
+                if (it_pcl == pcl_out.points.begin()) {
+                    break;
+                }
+            }
+        }
+
+    }
 
     PointCloudType::Ptr cur_pcl_un_;
     sensor_msgs::msg::Imu::ConstSharedPtr last_imu_;
@@ -69,294 +341,5 @@ class ImuProcess {
     bool imu_need_init_ = true;
 };
 
-ImuProcess::ImuProcess() : b_first_frame_(true), imu_need_init_(true) {
-    init_iter_num_ = 1;
-    Q_ = process_noise_cov();
-    cov_acc_ = common::V3D(0.1, 0.1, 0.1);
-    cov_gyr_ = common::V3D(0.1, 0.1, 0.1);
-    cov_bias_gyr_ = common::V3D(0.0001, 0.0001, 0.0001);
-    cov_bias_acc_ = common::V3D(0.0001, 0.0001, 0.0001);
-    mean_acc_ = common::V3D(0, 0, -1.0);
-    mean_gyr_ = common::V3D(0, 0, 0);
-    angvel_last_ = common::Zero3d;
-    Lidar_T_wrt_IMU_ = common::Zero3d;
-    Lidar_R_wrt_IMU_ = common::Eye3d;
-    last_imu_.reset(new sensor_msgs::msg::Imu());
-}
 
-ImuProcess::~ImuProcess() {}
-
-void ImuProcess::Reset() {
-    mean_acc_ = common::V3D(0, 0, -1.0);
-    mean_gyr_ = common::V3D(0, 0, 0);
-    angvel_last_ = common::Zero3d;
-    imu_need_init_ = true;
-    init_iter_num_ = 1;
-    v_imu_.clear();
-    IMUpose_.clear();
-    last_imu_.reset(new sensor_msgs::msg::Imu());
-    cur_pcl_un_.reset(new PointCloudType());
-}
-
-void ImuProcess::SetExtrinsic(const common::V3D &transl, const common::M3D &rot) {
-    Lidar_T_wrt_IMU_ = transl;
-    Lidar_R_wrt_IMU_ = rot;
-}
-
-Sophus::SE3d ImuProcess::GetExtrinsic(){
-    Sophus::SE3d extrinsicImuToLidar{ Lidar_R_wrt_IMU_, Lidar_T_wrt_IMU_ };
-    return extrinsicImuToLidar;
-}
-
-void ImuProcess::SetGyrCov(const common::V3D &scaler) { cov_gyr_scale_ = scaler; }
-
-void ImuProcess::SetAccCov(const common::V3D &scaler) { cov_acc_scale_ = scaler; }
-
-void ImuProcess::SetGyrBiasCov(const common::V3D &b_g) { cov_bias_gyr_ = b_g; }
-
-void ImuProcess::SetAccBiasCov(const common::V3D &b_a) { cov_bias_acc_ = b_a; }
-
-void ImuProcess::IMUInit(const common::MeasureGroup &meas, esekfom::esekf<state_ikfom, 12, input_ikfom> &kf_state,
-                         int &N) {
-    /** 1. initializing the gravity_, gyro bias, acc and gyro covariance
-     ** 2. normalize the acceleration measurenments to unit gravity_ **/
-
-    common::V3D cur_acc, cur_gyr;
-
-    if (b_first_frame_) {
-        Reset();
-        N = 1;
-        b_first_frame_ = false;
-        const auto &imu_acc = meas.imu_.front()->linear_acceleration;
-        const auto &gyr_acc = meas.imu_.front()->angular_velocity;
-        mean_acc_ << imu_acc.x, imu_acc.y, imu_acc.z;
-        mean_gyr_ << gyr_acc.x, gyr_acc.y, gyr_acc.z;
-    }
-
-    for (const auto &imu : meas.imu_) {
-        const auto &imu_acc = imu->linear_acceleration;
-        const auto &gyr_acc = imu->angular_velocity;
-        cur_acc << imu_acc.x, imu_acc.y, imu_acc.z;
-        cur_gyr << gyr_acc.x, gyr_acc.y, gyr_acc.z;
-
-        mean_acc_ += (cur_acc - mean_acc_) / N;
-        mean_gyr_ += (cur_gyr - mean_gyr_) / N;
-
-        cov_acc_ =
-            cov_acc_ * (N - 1.0) / N + (cur_acc - mean_acc_).cwiseProduct(cur_acc - mean_acc_) * (N - 1.0) / (N * N);
-        cov_gyr_ =
-            cov_gyr_ * (N - 1.0) / N + (cur_gyr - mean_gyr_).cwiseProduct(cur_gyr - mean_gyr_) * (N - 1.0) / (N * N);
-
-        N++;
-    }
-    state_ikfom init_state = kf_state.get_x();
-    init_state.grav = S2(-mean_acc_ / mean_acc_.norm() * common::G_m_s2);
-
-    init_state.bg = mean_gyr_;
-    init_state.offset_T_L_I = Lidar_T_wrt_IMU_;
-    init_state.offset_R_L_I = Lidar_R_wrt_IMU_;
-    kf_state.change_x(init_state);
-
-    esekfom::esekf<state_ikfom, 12, input_ikfom>::cov init_P = kf_state.get_P();
-    init_P.setIdentity();
-    init_P(6, 6) = init_P(7, 7) = init_P(8, 8) = 0.00001;
-    init_P(9, 9) = init_P(10, 10) = init_P(11, 11) = 0.00001;
-    init_P(15, 15) = init_P(16, 16) = init_P(17, 17) = 0.0001;
-    init_P(18, 18) = init_P(19, 19) = init_P(20, 20) = 0.001;
-    init_P(21, 21) = init_P(22, 22) = 0.00001;
-    kf_state.change_P(init_P);
-    last_imu_ = meas.imu_.back();
-}
-
-/**
- * @brief iekf forward propagation, use imu to get undistort lidar points in tail lidar_body frame
- * 
- * @param[in] meas imu dates and lidar points in once lidar scan
- * @param[in] kf_state pose state of iekf
- * @param[out] pcl_out undistort points output
- */
-void ImuProcess::UndistortPcl(const common::MeasureGroup &meas, esekfom::esekf<state_ikfom, 12, input_ikfom> &kf_state,
-                              PointCloudType &pcl_out) {
-    // Step 1.get head and tail time of imu and lidar
-    auto v_imu = meas.imu_;
-    v_imu.push_front(last_imu_);    // add the imu_ of the last frame-tail to the of current frame-head
-    const double &imu_beg_time = common::FromRosTime(v_imu.front()->header.stamp);
-    const double &imu_end_time = common::FromRosTime(v_imu.back()->header.stamp);
-    const double &pcl_beg_time = meas.lidar_bag_time_;
-    const double &pcl_end_time = meas.lidar_end_time_;
-
-    // Step 2.sort point clouds by offset time
-    pcl_out = *(meas.lidar_);
-    sort(pcl_out.points.begin(), pcl_out.points.end(), time_list);
-    // LOG(WARNING) << "points size: " << pcl_out.points.size();
-    // Step 3.Initialize tail IMU pose of last frame
-    state_ikfom imu_state = kf_state.get_x();
-    IMUpose_.clear();
-    IMUpose_.push_back(common::set_pose6d(0.0, acc_s_last_, angvel_last_, imu_state.vel, imu_state.pos,
-                                          imu_state.rot.toRotationMatrix()));
-
-    
-    common::V3D angvel_avr, acc_avr, acc_imu, vel_imu, pos_imu;
-    common::M3D R_imu;
-
-    double dt = 0;
-
-    input_ikfom in;
-    // Step 4.go through all imu vector， forward propagation at each imu_ date
-    for (auto it_imu = v_imu.begin(); it_imu < (v_imu.end() - 1); it_imu++) {
-        auto &&head = *(it_imu);
-        auto &&tail = *(it_imu + 1);
-        // Step 4.1 insure: last lidar end time << tail time
-        if ( common::FromRosTime(tail->header.stamp) < last_lidar_end_time_) {
-            continue;
-        }
-
-        // Step 4.2 get ave imu date
-        angvel_avr << 0.5 * (head->angular_velocity.x + tail->angular_velocity.x),
-            0.5 * (head->angular_velocity.y + tail->angular_velocity.y),
-            0.5 * (head->angular_velocity.z + tail->angular_velocity.z);
-        acc_avr << 0.5 * (head->linear_acceleration.x + tail->linear_acceleration.x),
-            0.5 * (head->linear_acceleration.y + tail->linear_acceleration.y),
-            0.5 * (head->linear_acceleration.z + tail->linear_acceleration.z);
-        
-        // Step 4.3 acc normalize to G_m_s2 length
-        acc_avr = acc_avr * common::G_m_s2 / mean_acc_.norm();  // - state_inout.ba;
-
-        // Step 4.4 get dt
-        if ( common::FromRosTime( head->header.stamp ) < last_lidar_end_time_) {
-            dt = common::FromRosTime( tail->header.stamp ) - last_lidar_end_time_;
-        } else {
-            dt = common::FromRosTime( tail->header.stamp ) - common::FromRosTime( head->header.stamp );
-        }
-
-        // Step 4.5 forward propagation this imu date
-        in.acc = acc_avr;
-        in.gyro = angvel_avr;
-        Q_.block<3, 3>(0, 0).diagonal() = cov_gyr_;
-        Q_.block<3, 3>(3, 3).diagonal() = cov_acc_;
-        Q_.block<3, 3>(6, 6).diagonal() = cov_bias_gyr_;
-        Q_.block<3, 3>(9, 9).diagonal() = cov_bias_acc_;
-        kf_state.predict(dt, Q_, in);
-
-        // Step 4.6 save the poses at each IMU measurements( in imu_world frame ) by Step4.5
-        imu_state = kf_state.get_x();
-        angvel_last_ = angvel_avr - imu_state.bg;
-        acc_s_last_ = imu_state.rot * (acc_avr - imu_state.ba);
-        for (int i = 0; i < 3; i++) {
-            acc_s_last_[i] += imu_state.grav[i];
-        }
-
-        double &&offs_t = common::FromRosTime( tail->header.stamp ) - pcl_beg_time;
-        // simple set off_t、acc、ang、vel、trans、rot
-        IMUpose_.emplace_back(common::set_pose6d(offs_t, acc_s_last_, angvel_last_, imu_state.vel, imu_state.pos,
-                                                 imu_state.rot.toRotationMatrix()));
-    }
-
-    // Step 5.forward to tail lidar point in cur frame
-    double note = pcl_end_time > imu_end_time ? 1.0 : -1.0;
-    dt = note * (pcl_end_time - imu_end_time);
-    kf_state.predict(dt, Q_, in);
-
-    imu_state = kf_state.get_x();
-    last_imu_ = meas.imu_.back();
-    last_lidar_end_time_ = pcl_end_time;
-
-    // Step 6.undistort each lidar point (backward propagation)
-    if (pcl_out.points.empty()) {
-        return;
-    }
-    auto it_pcl = pcl_out.points.end() - 1; // tail point
-    // LOG(WARNING) << "orgin itPcl: " << it_pcl->x << " " << it_pcl->y << " " << it_pcl->z;
-    // backward propagation imu pose
-    for (auto it_kp = IMUpose_.end() - 1; it_kp != IMUpose_.begin(); it_kp--) {
-        auto head = it_kp - 1;
-        auto tail = it_kp;
-        // R_imu = common::MatFromArray(head->rot);
-        // // LOG_FIRST_N(INFO,5) << "head->rot:" << std::endl << head->rot;
-        // // LOG_FIRST_N(INFO,5) << "R_imu:" << std::endl << R_imu;
-        // vel_imu = common::VecFromArray(head->vel);
-        // pos_imu = common::VecFromArray(head->pos);
-        // acc_imu = common::VecFromArray(tail->acc);
-        // angvel_avr = common::VecFromArray(tail->gyr);
-
-        R_imu = head->rot;
-        vel_imu = head->vel;
-        pos_imu = head->pos;
-        acc_imu = tail->acc;
-        angvel_avr = tail->gyr;
-
-        // LOG(INFO) << "imu pose vector offset time: " << head->offset_time;
-        // go through points between head and tail imu pose
-        for (; it_pcl->curvature / double(1000) > head->offset_time; it_pcl--) {
-            dt = it_pcl->curvature / double(1000) - head->offset_time;
-            // LOG(WARNING) << "dt: " << dt;
-            /* ! Transform to the 'end' frame, using only the rotation
-             * Note: Compensation direction is INVERSE of Frame's moving direction
-             * So if we want to compensate a point at timestamp-i to the frame-e
-             * p_compensate = R_imu_e ^ T * (R_i * P_i + T_ei) where T_ei is represented in global frame */
-            
-            // this point time rot in imu_world
-            common::M3D R_i(R_imu * Exp(angvel_avr, dt));
-            // this point time translate in lidar_body frame
-            common::V3D P_i(it_pcl->x, it_pcl->y, it_pcl->z);
-            // delta translate of (end time pose) to (this point time) in imu_world frame
-            common::V3D T_ei(pos_imu + vel_imu * dt + 0.5 * acc_imu * dt * dt - imu_state.pos);
-            // 
-            common::V3D p_compensate =
-                /*lidar point pose in tail lidar_body frame without rotation distort*/
-                imu_state.offset_R_L_I.conjugate() *
-                /*lidar point pose in tail imu_body frame without rotation distort*/
-                (imu_state.rot.conjugate() * 
-                    /*lidar point pose in imu_world frame with distort*/
-                    (R_i * 
-                        /*lidar point pose in imu_body frame*/
-                        (imu_state.offset_R_L_I * P_i + imu_state.offset_T_L_I)
-                    + T_ei) -
-                 imu_state.offset_T_L_I);  // not accurate!
-
-            // save Undistorted points and their rotation
-            it_pcl->x = p_compensate(0);
-            it_pcl->y = p_compensate(1);
-            it_pcl->z = p_compensate(2);
-            // LOG(WARNING) << "itPcl: " << it_pcl->x << " " << it_pcl->y << " " << it_pcl->z;
-            if (it_pcl == pcl_out.points.begin()) {
-                break;
-            }
-        }
-    }
-}
-
-void ImuProcess::Process(const common::MeasureGroup &meas, esekfom::esekf<state_ikfom, 12, input_ikfom> &kf_state,
-                         PointCloudType::Ptr cur_pcl_un_) {
-    if (meas.imu_.empty()) {
-        return;
-    }
-
-    if (imu_need_init_) {
-        /// The very first lidar frame
-        IMUInit(meas, kf_state, init_iter_num_);
-
-        imu_need_init_ = true;
-
-        last_imu_ = meas.imu_.back();
-
-        state_ikfom imu_state = kf_state.get_x();
-        if (init_iter_num_ > MAX_INI_COUNT) {
-            cov_acc_ *= pow(common::G_m_s2 / mean_acc_.norm(), 2);
-            imu_need_init_ = false;
-
-            cov_acc_ = cov_acc_scale_;
-            cov_gyr_ = cov_gyr_scale_;
-            LOG(WARNING) << "mean Acc: " << mean_acc_.transpose();
-            LOG(WARNING) << "cov Acc: " << cov_acc_.transpose();
-            LOG(WARNING) << "cov Gyro: " << cov_gyr_.transpose();
-            LOG(INFO) << "IMU Initial Done";
-            // fout_imu_.open(common::DEBUG_FILE_DIR("imu_.txt"), std::ios::out);  // ! maybe modify later
-        }
-
-        return;
-    }
-
-    Timer::Evaluate([&, this]() { UndistortPcl(meas, kf_state, *cur_pcl_un_); }, "Undistort Pcl");
-}
 }  // namespace fast_lio
